requirements.txt
asgiref>=3.8.1,<4
Django==5.1.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
sqlparse==0.4.4
tzdata==2024.1
numpy==1.26.2
pytz==2023.3.post1